## PR Description

Backports https://github.com/JuliaLang/julia/pull/59366.

------

In `--trace-compile` Tag nested precompiles with `# nested_compile`.

Before:
```julia
  ./usr/bin/julia --startup=no --trace-compile=stderr -e '
    Base.@assume_effects :foldable function nested1(x)
        sum(collect(x for _ in 1:10_000_000))
    end
    f1(x) = nested1(sizeof(x)) + x
    f1(2)'
  precompile(Tuple{typeof(Main.nested1), Int64})
  precompile(Tuple{typeof(Main.f1), Int64})

  ./usr/bin/julia --startup=no --trace-compile=stderr --trace-compile-timing -e '
    Base.@assume_effects :foldable function nested1(x)
        sum(collect(x for _ in 1:10_000_000))
    end
    f1(x) = nested1(sizeof(x)) + x
    f1(2)'
  #=    8.1 ms =# precompile(Tuple{typeof(Main.nested1), Int64})
  #=   71.3 ms =# precompile(Tuple{typeof(Main.f1), Int64})
```

After:
```julia
  ./usr/bin/julia --startup=no --trace-compile=stderr -e '
    Base.@assume_effects :foldable function nested1(x)
        sum(collect(x for _ in 1:10_000_000))
    end
    f1(x) = nested1(sizeof(x)) + x
    f1(2)'
  precompile(Tuple{typeof(Main.nested1), Int64}) # nested_const_compilation
  precompile(Tuple{typeof(Main.f1), Int64})

  ./usr/bin/julia --startup=no --trace-compile=stderr --trace-compile-timing -e '
    Base.@assume_effects :foldable function nested1(x)
        sum(collect(x for _ in 1:10_000_000))
    end
    f1(x) = nested1(sizeof(x)) + x
    f1(2)'
  #=    8.1 ms =# precompile(Tuple{typeof(Main.nested1), Int64}) # nested_const_compilation
  #=   71.3 ms =# precompile(Tuple{typeof(Main.f1), Int64})
```

## Checklist

Requirements for merging:
- [x] I have opened an issue or PR upstream on JuliaLang/julia: https://github.com/JuliaLang/julia/pull/59366
- [ ] I have removed the `port-to-*` labels that don't apply.
- [ ] I have opened a PR on raicode to test these changes: <link to raicode>